# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gavhah Community Kindness Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: List of news articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get news article by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update a news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /minyans:
    get:
      operationId: listMinyans
      tags: [minyans]
      summary: List minyan listings
      parameters:
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: community
          schema:
            type: string
        - in: query
          name: synagogue
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: List of minyans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Minyan"
    post:
      operationId: createMinyan
      tags: [minyans]
      summary: Submit a new minyan listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MinyanInput"
      responses:
        "201":
          description: Created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Minyan"
  /minyans/{id}:
    get:
      operationId: getMinyan
      tags: [minyans]
      summary: Get minyan by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Minyan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Minyan"
    patch:
      operationId: updateMinyan
      tags: [minyans]
      summary: Update a minyan listing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MinyanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Minyan"
  /minyans/{id}/like:
    post:
      operationId: likeMinyan
      tags: [minyans]
      summary: Like a minyan listing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

  # ---- VOLUNTEERS / HELP REQUESTS ----
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List platform announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- REPORTS ----
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List content reports (admin/moderator)
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a piece of content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/resolve:
    post:
      operationId: resolveReport
      tags: [reports]
      summary: Resolve a report (admin/moderator)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ---- NOTIFICATIONS ----
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    LikeResult:
      type: object
      required: [likes]
      properties:
        likes:
          type: integer

    # ---- NEWS ----
    NewsArticle:
      type: object
      required: [id, title, content, category, authorId, authorName, isFeatured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        summary:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
          enum: [medical, wedding, bikur_cholim, community, emergency, volunteer]
        authorId:
          type: integer
        authorName:
          type: string
        isFeatured:
          type: boolean
        commentCount:
          type: integer
        createdAt:
          type: string
    NewsArticleInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isFeatured:
          type: boolean
    NewsArticleUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        isFeatured:
          type: boolean

    # ---- GROUP ----
...
    Minyan:
      type: object
      required: [id, synagogueName, community, city, country, shacharis, mincha, maariv, likes, status, createdAt]
      properties:
        id:
          type: integer
        synagogueName:
          type: string
        community:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: ["string", "null"]
        shacharis:
          type: string
        mincha:
          type: string
        maariv:
          type: string
        notes:
          type: ["string", "null"]
        likes:
          type: integer
        status:
          type: string
          enum: [approved, pending, rejected]
        createdAt:
          type: string
    MinyanInput:
      type: object
      required: [synagogueName, community, city, country, shacharis, mincha, maariv]
      properties:
        synagogueName:
          type: string
        community:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        shacharis:
          type: string
        mincha:
          type: string
        maariv:
          type: string
        notes:
          type: string
    MinyanUpdate:
      type: object
      properties:
        synagogueName:
          type: string
        shacharis:
          type: string
        mincha:
          type: string
        maariv:
          type: string
        notes:
          type: string
        status:
          type: string

    # ---- VOLUNTEER ----
...
    Announcement:
      type: object
      required: [id, title, content, authorId, authorName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string

    # ---- REPORTS ----
    Report:
      type: object
      required: [id, contentType, contentId, reason, reporterId, status, createdAt]
      properties:
        id:
          type: integer
        contentType:
          type: string
          enum: [discussion, comment, group, news, minyan, help_request]
        contentId:
          type: integer
        reason:
          type: string
          enum: [spam, harassment, inappropriate, false_information, other]
        description:
          type: ["string", "null"]
        reporterId:
          type: integer
        status:
          type: string
          enum: [pending, resolved, dismissed]
        createdAt:
          type: string
    ReportInput:
      type: object
      required: [contentType, contentId, reason]
      properties:
        contentType:
          type: string
        contentId:
          type: integer
        reason:
          type: string
        description:
          type: ["string", "null"]

    # ---- NOTIFICATIONS ----
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [reply, group_activity, volunteer_match, announcement, charity_update]
        message:
          type: string
        isRead:
          type: boolean
        linkUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...